requirements.txt
streamlit  
python-dotenv
langchain
langchain-google-genai
langgraph
pydantic
pymongo
